+++ packages/blaze-emulator/src/emulator.ts
@@ -137,7 +137,18 @@ export class Emulator {
     }
     this.clock = new LedgerTimer(slotConfig);
     this.params = params;
-    this.evaluator = evaluator ?? makeUplcEvaluator(params, 1, 1, slotConfig);
+    this.evaluator =
+      evaluator ??
+      makeUplcEvaluator(
+        params,
+        1,
+        1,
+        slotConfig ?? {
+          zeroSlot: this.clock.slot,
+          zeroTime: this.clock.time,
+          slotLength: this.clock.slotLength,
+        },
+      );
     this.addUtxo = this.addUtxo.bind(this);
     this.removeUtxo = this.removeUtxo.bind(this);
   }
@@ -493,7 +504,7 @@ export class Emulator {
 
     // Minimum collateral amount included
     const minCollateral = BigInt(
-      Math.ceil(this.params.collateralPercentage * Number(body.fee())),
+      Math.ceil((this.params.collateralPercentage * Number(body.fee())) / 100),
     );
 
     // If any scripts have been invoked, minimum collateral must be included