+++ packages/blaze-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blaze-cardano/core
 
+## 0.4.5
+
+### Patch Changes
+
+- 3d0f4b7: Fix issue with blakejs import
+
+## 0.4.4
+
+### Patch Changes
+
+- 476d4af: Add a custom UTXO selection error
+
 ## 0.4.3
 
 ### Patch Changes