+++ packages/blaze-emulator/CHANGELOG.md
@@ -1,5 +1,112 @@
 # @blaze-cardano/emulator
 
+## 0.1.75
+
+### Patch Changes
+
+- Updated dependencies [c924f95]
+- Updated dependencies [471d47d]
+- Updated dependencies [c951eec]
+  - @blaze-cardano/tx@0.6.0
+  - @blaze-cardano/wallet@0.1.55
+
+## 0.1.74
+
+### Patch Changes
+
+- Updated dependencies [bdf17aa]
+- Updated dependencies [73e34a6]
+  - @blaze-cardano/tx@0.5.17
+  - @blaze-cardano/wallet@0.1.54
+
+## 0.1.73
+
+### Patch Changes
+
+- Updated dependencies [ff8fd0a]
+  - @blaze-cardano/tx@0.5.16
+  - @blaze-cardano/wallet@0.1.53
+
+## 0.1.72
+
+### Patch Changes
+
+- @blaze-cardano/query@0.2.17
+- @blaze-cardano/wallet@0.1.52
+
+## 0.1.71
+
+### Patch Changes
+
+- Updated dependencies [4cddb94]
+  - @blaze-cardano/tx@0.5.15
+  - @blaze-cardano/wallet@0.1.51
+
+## 0.1.70
+
+### Patch Changes
+
+- Updated dependencies [3d0f4b7]
+  - @blaze-cardano/core@0.4.5
+  - @blaze-cardano/query@0.2.16
+  - @blaze-cardano/tx@0.5.14
+  - @blaze-cardano/vm@0.0.38
+  - @blaze-cardano/wallet@0.1.50
+
+## 0.1.69
+
+### Patch Changes
+
+- Updated dependencies [e12d65e]
+  - @blaze-cardano/tx@0.5.13
+  - @blaze-cardano/wallet@0.1.49
+
+## 0.1.68
+
+### Patch Changes
+
+- Updated dependencies [bb5da39]
+  - @blaze-cardano/query@0.2.15
+  - @blaze-cardano/tx@0.5.12
+  - @blaze-cardano/wallet@0.1.48
+
+## 0.1.67
+
+### Patch Changes
+
+- Updated dependencies [5e2b970]
+  - @blaze-cardano/tx@0.5.11
+  - @blaze-cardano/wallet@0.1.47
+
+## 0.1.66
+
+### Patch Changes
+
+- Updated dependencies [5369627]
+  - @blaze-cardano/tx@0.5.10
+  - @blaze-cardano/wallet@0.1.46
+
+## 0.1.65
+
+### Patch Changes
+
+- Updated dependencies [dcb0d76]
+  - @blaze-cardano/tx@0.5.9
+  - @blaze-cardano/wallet@0.1.45
+
+## 0.1.64
+
+### Patch Changes
+
+- Updated dependencies [476d4af]
+- Updated dependencies [3946a63]
+- Updated dependencies [a4e6808]
+  - @blaze-cardano/core@0.4.4
+  - @blaze-cardano/tx@0.5.8
+  - @blaze-cardano/query@0.2.14
+  - @blaze-cardano/vm@0.0.37
+  - @blaze-cardano/wallet@0.1.44
+
 ## 0.1.63
 
 ### Patch Changes